Modify tests
