+++ test/fixtures/bem-tools/getTechnologies.bemjson.json
@@ -1,9 +1,9 @@
 {
   "inBlocks": {
     "V2": [
-      "'bemdecl.js'           : 'v2/bemdecl.js'",
-      "'deps.js'              : 'v2/deps.js'",
-      "'css'                  : 'v2/css'"
+      "'bemdecl.js'           : 'bemdecl.js'",
+      "'deps.js'              : 'deps.js'",
+      "'css'                  : 'css'"
     ],
     "notV2": [],
     "defaultTechs": ["css"]
