+++ test/fixtures/bem-tools/getTechnologies.ie.json
@@ -1,14 +1,14 @@
 {
   "inBlocks": {
     "V2": [
-      "'bemdecl.js'           : 'v2/bemdecl.js'",
-      "'deps.js'              : 'v2/deps.js'",
-      "'css'                  : 'v2/css'",
-      "'ie.css'               : 'v2/ie.css'",
-      "'ie6.css'              : 'v2/ie6.css'",
-      "'ie7.css'              : 'v2/ie7.css'",
-      "'ie8.css'              : 'v2/ie8.css'",
-      "'ie9.css'              : 'v2/ie9.css'"
+      "'bemdecl.js'           : 'bemdecl.js'",
+      "'deps.js'              : 'deps.js'",
+      "'css'                  : 'css'",
+      "'ie.css'               : 'ie.css'",
+      "'ie6.css'              : 'ie6.css'",
+      "'ie7.css'              : 'ie7.css'",
+      "'ie8.css'              : 'ie8.css'",
+      "'ie9.css'              : 'ie9.css'"
     ],
     "notV2": [],
     "defaultTechs": ["css"]
