+++ test/fixtures/bem-tools/getTechnologies.less.json
@@ -1,10 +1,10 @@
 {
   "inBlocks": {
     "V2": [
-      "'bemdecl.js'           : 'v2/bemdecl.js'",
-      "'deps.js'              : 'v2/deps.js'",
-      "'less'                 : 'v2/less'",
-      "'css'                  : 'v2/css'"
+      "'bemdecl.js'           : 'bemdecl.js'",
+      "'deps.js'              : 'deps.js'",
+      "'less'                 : 'less'",
+      "'css'                  : 'css'"
     ],
     "notV2": [],
     "defaultTechs": ["less"]
