+++ test/fixtures/bem-tools/getTechnologies.stylus+autoprefixer.json
@@ -1,10 +1,10 @@
 {
   "inBlocks": {
     "V2": [
-      "'bemdecl.js'           : 'v2/bemdecl.js'",
-      "'deps.js'              : 'v2/deps.js'",
-      "'stylus'               : 'v2/styl'",
-      "'css'                  : 'v2/css'"
+      "'bemdecl.js'           : 'bemdecl.js'",
+      "'deps.js'              : 'deps.js'",
+      "'stylus'               : 'styl'",
+      "'css'                  : 'css'"
     ],
     "notV2": [],
     "defaultTechs": ["stylus"]
