+++ test/fixtures/enb/getTechnologies.bemtree.json
@@ -1,9 +1,7 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb-bemxjst/techs/bemtree-old')",
-    "require('enb/techs/css')"
+    "bemtree",
+    "css"
   ],
   "inTargets": ["?.css", "?.bemtree.js"],
   "inJSON": [{
@@ -14,6 +12,9 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
     "version": "*"
   }, {
     "name": "enb-bemxjst",
