+++ test/fixtures/enb/getTechnologies.bh.json
@@ -1,9 +1,7 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb-bh/techs/bh-server')",
-    "require('enb/techs/css')"
+    "bh",
+    "css"
   ],
   "inTargets": ["?.css", "?.bh.js"],
   "inJSON": [{
@@ -14,6 +12,9 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
     "version": "*"
   }, {
     "name": "enb-bh",
