+++ test/fixtures/enb/getTechnologies.bh.min.json
@@ -1,9 +1,7 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb-bh/techs/bh-server')",
-    "require('enb/techs/css')"
+    "bh",
+    "css"
   ],
   "inTargets": ["?.css", "_?.bh.js"],
   "inJSON": [{
@@ -14,6 +12,12 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
+    "version": "*"
+  }, {
+    "name": "enb-borschik",
     "version": "*"
   }, {
     "name": "enb-bh",
