+++ test/fixtures/enb/getTechnologies.css+autoprefixer.min.json
@@ -1,8 +1,6 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb/techs/css'), { target: '?.noprefix.css' }"
+    "css+autoprefixer"
   ],
   "inTargets": ["_?.css"],
   "inJSON": [{
@@ -15,7 +13,16 @@
     "name": "enb",
     "version": "*"
   }, {
+    "name": "enb-bem-techs",
+    "version": "*"
+  }, {
     "name": "enb-autoprefixer",
+    "version": "*"
+  }, {
+    "name": "enb-borschik",
+    "version": "*"
+  }, {
+    "name": "borschik-tech-cleancss",
     "version": "*"
   }]
 }
