+++ test/fixtures/enb/getTechnologies.css.json
@@ -1,8 +1,6 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb/techs/css')"
+    "css"
   ],
   "inTargets": ["?.css"],
   "inJSON": [{
@@ -13,6 +11,9 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
     "version": "*"
   }]
 }
