+++ test/fixtures/enb/getTechnologies.css.min.json
@@ -1,8 +1,6 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb/techs/css')"
+    "css"
   ],
   "inTargets": ["_?.css"],
   "inJSON": [{
@@ -13,6 +11,15 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
+    "version": "*"
+  }, {
+    "name": "enb-borschik",
+    "version": "*"
+  }, {
+    "name": "borschik-tech-cleancss",
     "version": "*"
   }]
 }
