+++ test/fixtures/enb/getTechnologies.html.browser+bemhtml.json
@@ -1,8 +1,13 @@
 {
   "inTechs": [
-    "stylus+autoprefixer"
+    "bemjson.js",
+    "css",
+    "browser.bemhtml.js",
+    "bemhtml",
+    "bemhtml-client",
+    "html-from-bemhtml"
   ],
-  "inTargets": ["?.css"],
+  "inTargets": ["?.css", "?.js", "?.bemhtml.js", "?.html"],
   "inJSON": [{
     "name": "bower",
     "version": "*"
@@ -16,10 +21,13 @@
     "name": "enb-bem-techs",
     "version": "*"
   }, {
-    "name": "enb-autoprefixer",
+    "name": "enb-diverse-js",
     "version": "*"
   }, {
-    "name": "enb-stylus",
+    "name": "enb-modules",
+    "version": "*"
+  }, {
+    "name": "enb-bemxjst",
     "version": "*"
   }]
 }
