+++ test/fixtures/enb/getTechnologies.ie.json
@@ -1,13 +1,11 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb/techs/css'), { target: '?.ie.css', sourceSuffixes: ['css', 'ie.css'] }",
-    "require('enb/techs/css'), { target: '?.ie6.css', sourceSuffixes: ['css', 'ie6.css'] }",
-    "require('enb/techs/css'), { target: '?.ie7.css', sourceSuffixes: ['css', 'ie7.css'] }",
-    "require('enb/techs/css'), { target: '?.ie8.css', sourceSuffixes: ['css', 'ie8.css'] }",
-    "require('enb/techs/css'), { target: '?.ie9.css', sourceSuffixes: ['css', 'ie9.css'] }",
-    "require('enb/techs/css')"
+    "ie.css",
+    "ie6.css",
+    "ie7.css",
+    "ie8.css",
+    "ie9.css",
+    "css"
   ],
   "inTargets": ["?.css", "?.ie.css", "?.ie6.css", "?.ie7.css", "?.ie8.css", "?.ie9.css"],
   "inJSON": [{
@@ -18,6 +16,9 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
     "version": "*"
   }]
 }
