+++ test/fixtures/enb/getTechnologies.less.json
@@ -1,8 +1,6 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb/techs/css-less')"
+    "less"
   ],
   "inTargets": ["?.css"],
   "inJSON": [{
@@ -13,6 +11,9 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
     "version": "*"
   }, {
     "name": "less",
