+++ test/fixtures/enb/getTechnologies.node.min.json
@@ -1,10 +1,7 @@
 {
   "inTechs": [
-    "require('enb/techs/files')",
-    "require('enb/techs/deps')",
-    "require('enb-diverse-js/techs/node-js'), { target: '?.pre.node.js' }",
-    "require('enb-modules/techs/prepend-modules'), { source: '?.pre.node.js', target: '?.node.js' }",
-    "require('enb/techs/css')"
+    "node.js",
+    "css"
   ],
   "inTargets": ["?.css", "_?.node.js"],
   "inJSON": [{
@@ -15,6 +12,12 @@
     "version": "*"
   }, {
     "name": "enb",
+    "version": "*"
+  }, {
+    "name": "enb-bem-techs",
+    "version": "*"
+  }, {
+    "name": "enb-borschik",
     "version": "*"
   }, {
     "name": "enb-diverse-js",
