+++ test/mocks/browsers.js
@@ -1,0 +1,17 @@
+module.exports = {
+    desktop: [
+        'last 2 versions',
+        'ie 10',
+        'ff 24',
+        'opera 12.16'
+    ],
+    'touch-pad': [
+        'android 4',
+        'ios 5'
+    ],
+    'touch-phone': [
+        'android 4',
+        'ios 6',
+        'ie 10'
+    ]
+};
