+++ test/mocks/techs.js
@@ -1,0 +1,42 @@
+module.exports = {
+    enb: {
+        'bemjson.js': 'bemjson.js',
+        roole: 'roole',
+        'roole+autoprefixer': 'roole+autoprefixer',
+        less: 'less',
+        'less+autoprefixer': 'less+autoprefixer',
+        stylus: 'stylus',
+        'stylus+autoprefixer': 'stylus+autoprefixer',
+        css: 'css',
+        'css+autoprefixer': 'css+autoprefixer',
+        'ie.css': 'ie.css',
+        'ie6.css': 'ie6.css',
+        'ie7.css': 'ie7.css',
+        'ie8.css': 'ie8.css',
+        'ie9.css': 'ie9.css',
+        bemtree: 'bemtree',
+        bemhtml: 'bemhtml',
+        'bemhtml-client': 'bemhtml-client',
+        bh: 'bh',
+        'bh-client': 'bh-client',
+        'node.js': 'node.js',
+        'browser.js': 'browser.js',
+        'browser.js+template': 'browser.template.js',
+        'html-from-bemhtml': 'html-from-bemhtml',
+        'html-from-bh': 'html-from-bh'
+    },
+    'bem-tools': {
+        'bemdecl.js': 'bemdecl.js',
+        'deps.js': 'deps.js',
+        stylus: 'styl',
+        roole: 'roole',
+        less: 'less',
+        css: 'css',
+        'ie.css': 'ie.css',
+        'ie6.css': 'ie6.css',
+        'ie7.css': 'ie7.css',
+        'ie8.css': 'ie8.css',
+        'ie9.css': 'ie9.css',
+        js: 'js-i'
+    }
+};
