+++ test/mocks/versions.js
@@ -1,0 +1,28 @@
+module.exports = {
+    libs: {
+        'bem-core': '*',
+        'bem-components': '*'
+    },
+    deps: {
+        enb: '*',
+        'enb-autoprefixer': '*',
+        'enb-bemxjst': '*',
+        'enb-bem-techs': '*',
+        'enb-bh': '*',
+        'enb-borschik': '*',
+        'enb-diverse-js': '*',
+        'enb-modules': '*',
+        'enb-roole': '*',
+        'enb-stylus': '*',
+        bem: '*',
+        'bem-environ': '*',
+        'bem-tools-autoprefixer': '*',
+        bh: '*',
+        bower: '*',
+        'borschik-tech-cleancss': '*',
+        'bower-npm-install': '*',
+        less: '*',
+        roole: '*',
+        stylus: '*'
+    }
+};
