--- test/test.config.json
@@ -1,68 +1,0 @@
-{
-  "versions": {
-    "libs": {
-      "bem-core": "*",
-      "bem-components": "*"
-    },
-    "deps": {
-      "bower": "*",
-      "bower-npm-install": "*",
-      "stylus": "*",
-      "roole": "*",
-      "less": "*",
-      "bh": "*",
-      "enb": "*",
-      "enb-roole": "*",
-      "enb-stylus": "*",
-      "enb-diverse-js": "*",
-      "enb-bemxjst": "*",
-      "enb-modules": "*",
-      "enb-bh": "*",
-      "enb-autoprefixer": "*",
-      "bem": "*",
-      "bem-environ": "*",
-      "bem-tools-autoprefixer": "*"
-    }
-  },
-  "techs": {
-    "bem-tools": {
-      "bemdecl.js"         : "v2/bemdecl.js",
-      "deps.js"            : "v2/deps.js",
-      "stylus"             : "v2/styl",
-      "roole"              : "v2/roole",
-      "less"               : "v2/less",
-      "css"                : "v2/css",
-      "ie.css"             : "v2/ie.css",
-      "ie6.css"            : "v2/ie6.css",
-      "ie7.css"            : "v2/ie7.css",
-      "ie8.css"            : "v2/ie8.css",
-      "ie9.css"            : "v2/ie9.css",
-      "js"                 : "v2/js-i"
-    },
-    "enb": {
-      "bemjson.js"         : "require('enb/techs/bemdecl-from-bemjson')",
-      "roole"              : "require('enb-roole/techs/css-roole')",
-      "less"               : "require('enb/techs/css-less')",
-      "stylus"             : "require('enb-stylus/techs/css-stylus')",
-      "css"                : "require('enb/techs/css')",
-      "ie.css"             : "require('enb/techs/css'), { target: '?.ie.css', sourceSuffixes: ['css', 'ie.css'] }",
-      "ie6.css"            : "require('enb/techs/css'), { target: '?.ie6.css', sourceSuffixes: ['css', 'ie6.css'] }",
-      "ie7.css"            : "require('enb/techs/css'), { target: '?.ie7.css', sourceSuffixes: ['css', 'ie7.css'] }",
-      "ie8.css"            : "require('enb/techs/css'), { target: '?.ie8.css', sourceSuffixes: ['css', 'ie8.css'] }",
-      "ie9.css"            : "require('enb/techs/css'), { target: '?.ie9.css', sourceSuffixes: ['css', 'ie9.css'] }",
-      "bemtree"            : "require('enb-bemxjst/techs/bemtree-old')",
-      "core-bemhtml"       : "require('enb-bemxjst/techs/bemhtml-old')",
-      "bh"                 : "require('enb-bh/techs/bh-server')",
-      "pre-node.js"        : "require('enb-diverse-js/techs/node-js'), { target: '?.pre.node.js' }",
-      "node.js"            : "require('enb-modules/techs/prepend-modules'), { source: '?.pre.node.js', target: '?.node.js' }",
-      "pre-browser.js"     : "require('enb-diverse-js/techs/browser-js'), { target: '?.pre.js' }",
-      "browser.js"         : "require('enb-modules/techs/prepend-modules'), { source: '?.pre.js', target: '?.js' }",
-      "html"               : "require('enb/techs/html-from-bemjson')"
-    }
-  },
-  "browsers": {
-    "desktop": ["last 2 versions", "ie 10", "ff 24", "opera 12.16"],
-    "touch-pad": ["android 4", "ios 5"],
-    "touch-phone": ["android 4", "ios 6", "ie 10"]
-  }
-}
